+++ skeleton/settings_qa_mid_site.py
@@ -1,5 +1,0 @@
-from skeleton.settings_qa_base import *
-
-
-FOUNDRY['layers'] = ('mid', 'basic',)
-SITE_ID = 3